requirements.txt
groq>=0.9.0
click>=8.1.0
rich>=13.7.0
python-dotenv>=1.0.0
gitpython>=3.1.40
tree-sitter>=0.21.0
pytest>=8.0.0
requests>=2.31.0

